move things from api_client.models to more appropriate places
