+++ qiskit_aqt_provider/api_client/models_direct.py
@@ -12,11 +12,22 @@
 
 """API models specific to the direct access API."""
 
+import importlib.metadata
+import platform
 import uuid
-from typing import Annotated, Literal, Union
+from typing import Annotated, Final, Literal, Optional, Union
 
+import httpx
 import pydantic as pdt
 from typing_extensions import Self
+
+PACKAGE_VERSION: Final = importlib.metadata.version("qiskit-aqt-provider")
+USER_AGENT: Final = " ".join(
+    [
+        f"aqt-api-client/{PACKAGE_VERSION}",
+        f"({platform.system()}; {platform.python_implementation()}/{platform.python_version()})",
+    ]
+)
 
 
 class NumIons(pdt.BaseModel):
@@ -63,3 +74,21 @@ class JobResult(pdt.BaseModel):
             result: mock measured samples.
         """
         return cls(job_id=job_id, payload=JobResultFinished(result=result))
+
+
+def http_client(
+    *, base_url: str, token: str, user_agent_extra: Optional[str] = None
+) -> httpx.Client:
+    """A pre-configured httpx Client.
+
+    Args:
+        base_url: base URL of the server
+        token: access token for the remote service.
+        user_agent_extra: optional extra data to add to the user-agent string.
+    """
+    user_agent_extra = f" {user_agent_extra}" if user_agent_extra else ""
+    headers = {"User-Agent": USER_AGENT + user_agent_extra}
+    if token:
+        headers["Authorization"] = f"Bearer {token}"
+
+    return httpx.Client(headers=headers, base_url=base_url, timeout=10.0, follow_redirects=True)
